% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{add_opts}
\alias{add_opts}
\title{Save options()$gpphelper to system file}
\usage{
add_opts(opts = options()$gpphelper)
}
\arguments{
\item{opts}{options()$gpphelper}
}
\description{
Save options()$gpphelper to system file
}
\keyword{internal}
