% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_insert.R
\name{help_insert_comment}
\alias{help_insert_comment}
\title{insert template comment before current line}
\usage{
help_insert_comment(.comment_config = options()$gpphelper$commentTemplate)
}
\arguments{
\item{.comment_config}{a file path to import the template comment}
}
\description{
insert template comment before current line
}
\keyword{internal}
