% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_insert.R
\name{help_insert_header}
\alias{help_insert_header}
\title{insert template header before current program}
\usage{
help_insert_header(opts = options()$gpphelper)
}
\arguments{
\item{.header_config}{a file path to import the template header}
}
\description{
insert template header before current program
}
\keyword{internal}
