% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_insert.R
\name{reformat_row}
\alias{reformat_row}
\title{reformat the line of template by keywords}
\usage{
reformat_row(template, row_item, add_item)
}
\arguments{
\item{template}{the template read from .txt}

\item{row_item}{the keyword need to be updated}

\item{add_item}{the text need to be added in the end of row_item}
}
\description{
reformat the line of template by keywords
}
\keyword{internal}
